+++ dry_monads/do_notation.py
@@ -6,6 +6,9 @@ from typing import Callable
 from dry_monads.primitives.exceptions import UnwrapFailedError
 from dry_monads.primitives.types import MonadType
 
+# Typing decorators is not an easy task, see:
+# https://github.com/python/mypy/issues/3157
+
 
 def do_notation(
     function: Callable[..., MonadType],
@@ -13,16 +16,7 @@ def do_notation(
     """
     Decorator to enable 'do-notation' context.
 
-    See example usages below.
-
-    .. code:: python
-
-
-    See also:
-        - https://dry-rb.org/gems/dry-monads/do-notation/
-        - https://en.wikibooks.org/wiki/Haskell/do_notation
-        - https://wiki.haskell.org/Do_notation_considered_harmful
-
+    Should be used for series of computations that rely on ``.unwrap`` method.
     """
     @wraps(function)
     def decorator(*args, **kwargs) -> MonadType: