+++ tests/test_either/test_either_unwrap.py
@@ -7,11 +7,11 @@ from dry_monads.primitives.exceptions import UnwrapFailedError
 
 
 def test_unwrap_success():
-    """Ensures that left identity works for Right monad."""
+    """Ensures that unwrap works for Right monad."""
     assert Right(5).unwrap() == 5
 
 
 def test_unwrap_failure():
-    """Ensures that left identity works for Right monad."""
+    """Ensures that unwrap works for Left monad."""
     with pytest.raises(UnwrapFailedError):
         assert Left(5).unwrap()